feat: add validations and adjust customer schema

+++ src/services/customer.service.js
@@ -9,6 +9,11 @@ class CustomerService {
     }
 
     async getCustomerInfo(customerEmail) {
+
+        if (!customerEmail) {
+            throw new Error('Customer is required');
+        }
+
         try {
 
             const customer = await this.customerRepository.getCustomerInfo(customerEmail);
@@ -25,10 +30,17 @@ class CustomerService {
     async createCustomer(customer) {
 
         if (!customer) {
-            throw new Error('Customer is required');
+            throw new Error('Customer data is required');
+        }
+
+        if (!customer.password) {
+            throw new Error('Password is required');
         }
 
         try {
+
+            customer.password = await bcrypt.hash(customer.password, 10);
+
             return await this.customerRepository.createCustomer(customer);
         } catch (error) {
             throw new Error('Error creating customer');
@@ -36,7 +48,19 @@ class CustomerService {
     }
 
     async updateCustomer(customerEmail, customerData) {
+
+        if(!customerEmail) {
+            throw new Error('Email is required');
+        }
+        if(!customerData) {
+            throw new Error('Customer data is required');
+        }
+
         try {
+
+         if('password' in customerData) {
+            throw new Error('Password cannot be updated');
+         }
             return await this.customerRepository.updateCustomer(customerEmail, customerData);
         } catch (error) {
             throw new Error('Error updating customer: ' + error.message);
@@ -45,6 +69,10 @@ class CustomerService {
     
 
     async deleteCustomer(customerEmail) {
+        if(!customerEmail) {
+            throw new Error('Customer email is required');
+        }
+
         try {
             return await this.customerRepository.deleteCustomer(customerEmail);
         } catch (error) {
